+++ src/xrCore/crc32.cpp
@@ -44,7 +44,7 @@ constexpr std::array<u32, 256> generate_crc32_lookup_table() noexcept
 // Lookup table array
 static constexpr auto crc32_table = generate_crc32_lookup_table();
 
-u32 crc32(const void* P, u32 len)
+u32 crc32(const void* P, size_t len)
 {
     // Pass a text string to this function and it will return the CRC.
 
@@ -68,7 +68,7 @@ u32 crc32(const void* P, u32 len)
     return ulCRC ^ 0xffffffff;
 }
 
-u32 crc32(const void* P, u32 len, u32 starting_crc)
+u32 crc32(const void* P, size_t len, u32 starting_crc)
 {
     u32 ulCRC = 0xffffffff ^ starting_crc;
     u8* buffer = (u8*)P;
@@ -79,7 +79,7 @@ u32 crc32(const void* P, u32 len, u32 starting_crc)
     return ulCRC ^ 0xffffffff;
 }
 
-u32 path_crc32(const char* path, u32 len)
+u32 path_crc32(const char* path, size_t len)
 {
     u32 ulCRC = 0xffffffff;
     u8* buffer = (u8*)path;
