+++ src/xrCore/log.cpp
@@ -81,7 +81,7 @@ void Log(pcstr s)
 {
     int i, j;
 
-    const u32 length = xr_strlen(s);
+    const size_t length = xr_strlen(s);
     pstr split = static_cast<pstr>(xr_alloca((length + 1) * sizeof(char)));
     for (i = 0, j = 0; s[i] != 0; i++)
     {
@@ -125,7 +125,7 @@ void Log(pcstr msg, pcstr dop)
         return;
     }
 
-    const u32 buffer_size = (xr_strlen(msg) + 1 + xr_strlen(dop) + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 1 + xr_strlen(dop) + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
     strconcat(buffer_size, buf, msg, " ", dop);
     Log(buf);
@@ -133,7 +133,7 @@ void Log(pcstr msg, pcstr dop)
 
 void Log(pcstr msg, int dop)
 {
-    const u32 buffer_size = (xr_strlen(msg) + 1 + 11 + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 1 + 11 + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
 
     xr_sprintf(buf, buffer_size, "%s %i", msg, dop);
@@ -142,7 +142,7 @@ void Log(pcstr msg, int dop)
 
 void Log(pcstr msg, unsigned int dop)
 {
-    const u32 buffer_size = (xr_strlen(msg) + 1 + 10 + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 1 + 10 + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
 
     xr_sprintf(buf, buffer_size, "%s %u", msg, dop);
@@ -151,7 +151,7 @@ void Log(pcstr msg, unsigned int dop)
 
 void Log(pcstr msg, long dop)
 {
-    const u32 buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
 
     xr_sprintf(buf, buffer_size, "%s %li", msg, dop);
@@ -160,7 +160,7 @@ void Log(pcstr msg, long dop)
 
 void Log(pcstr msg, unsigned long dop)
 {
-    const u32 buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
 
     xr_sprintf(buf, buffer_size, "%s %lu", msg, dop);
@@ -169,7 +169,7 @@ void Log(pcstr msg, unsigned long dop)
 
 void Log(pcstr msg, long long dop)
 {
-    const u32 buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
 
     xr_sprintf(buf, buffer_size, "%s %lli", msg, dop);
@@ -178,7 +178,7 @@ void Log(pcstr msg, long long dop)
 
 void Log(pcstr msg, unsigned long long dop)
 {
-    const u32 buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
 
     xr_sprintf(buf, buffer_size, "%s %llu", msg, dop);
@@ -189,7 +189,7 @@ void Log(pcstr msg, float dop)
 {
     // actually, float string representation should be no more, than 40 characters,
     // but we will count with slight overhead
-    const u32 buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 1 + 64 + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
 
     xr_sprintf(buf, buffer_size, "%s %f", msg, dop);
@@ -198,7 +198,7 @@ void Log(pcstr msg, float dop)
 
 void Log(pcstr msg, const Fvector& dop)
 {
-    const u32 buffer_size = (xr_strlen(msg) + 2 + 3 * (64 + 1) + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 2 + 3 * (64 + 1) + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
 
     xr_sprintf(buf, buffer_size, "%s (%f,%f,%f)", msg, VPUSH(dop));
@@ -207,7 +207,7 @@ void Log(pcstr msg, const Fvector& dop)
 
 void Log(pcstr msg, const Fmatrix& dop)
 {
-    const u32 buffer_size = (xr_strlen(msg) + 2 + 4 * (4 * (64 + 1) + 1) + 1) * sizeof(char);
+    const size_t buffer_size = (xr_strlen(msg) + 2 + 4 * (4 * (64 + 1) + 1) + 1) * sizeof(char);
     pstr buf = static_cast<pstr>(xr_alloca(buffer_size));
 
     xr_sprintf(buf, buffer_size, "%s:\n%f,%f,%f,%f\n%f,%f,%f,%f\n%f,%f,%f,%f\n%f,%f,%f,%f\n", msg, dop.i.x, dop.i.y,
